+++ server/src/main/java/com/ep/mqtt/server/db/dto/ReceiveMessageDto.java
@@ -1,11 +1,10 @@
 package com.ep.mqtt.server.db.dto;
 
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.ep.mqtt.server.metadata.Qos;
 import com.ep.mqtt.server.metadata.YesOrNo;
+
 import lombok.Data;
 
 /**
@@ -36,6 +35,6 @@ public class ReceiveMessageDto {
 
     private YesOrNo isComplete;
 
-    private Integer validTime;
+    private Long validTime;
 
 }
